+++ src/components/Setup/Setup.js
@@ -5,6 +5,7 @@ import Step from '@material-ui/core/Step';
 import StepLabel from '@material-ui/core/StepLabel';
 import StepContent from '@material-ui/core/StepContent';
 import Button from '@material-ui/core/Button';
+import CloudUploadIcon from '@material-ui/icons/CardTravel';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import TextField from '@material-ui/core/TextField';
@@ -13,7 +14,7 @@ import Lab from '../Lab/Lab';
 import bg from '../../assets/images/bg-new.png';
 import './setup.css';
 import {UserContext} from '../../store/context/userContext';
-
+import OptionsButton from '../OptionsButton'
 const useStyles = makeStyles(theme => ({
   root: {
     width: '100%',
@@ -47,7 +48,13 @@ const useStyles = makeStyles(theme => ({
   },
   iconContainer : {
     fontStyle:'italic'
-  }
+  },
+  button: {
+    margin: theme.spacing(1),
+  },
+  rightIcon: {
+    marginLeft: theme.spacing(1),
+  },
 }));
 
 function getSteps() {
@@ -166,12 +173,19 @@ function Setup(props) {
           <br/> "Worries end when Salah begins"
                       {/* <span onClick={handleTravel} style={{fontWeight:'bold', cursor:'pointer'}}>Click here...</span> (Alpha Release) */}
         </Typography> }
-        <span onClick={() => setPage('Travel')}>
-        <Typography color="textSecondary" variant="h2" component="h2" align="left" 
+        {/* <span onClick={() => setPage('Travel')}> */}
+        {/* <Typography color="textSecondary" variant="h2" component="h2" align="left" 
       style={{fontWeight:'bold', fontSize:'1rem', padding:'10px 24px', color: 'rgb(3, 155, 229)', marginBottom:0}} gutterBottom>
                   Travellers click here ... ( in α state)
-      </Typography>
-      </span>
+      </Typography> */}
+      {/* <Button variant="contained" color="primary" className={classes.button} onClick={() => setPage('Travel')} align="left" style={{color:'#fff'}}>
+        Travellers */}
+        {/* This Button uses a Font Icon, see the installation instructions in the docs. */}
+        {/* <CloudUploadIcon className={classes.rightIcon} />
+      </Button> */}
+      <OptionsButton options={[{opt : 'Traveller Onboard', page : 'Travel'}, {opt : 'Near by Mosques', page : ''}, {opt : 'Maps', page : ''}]}
+       setPage={(pge) => setPage(pge)} />
+      {/* </span> */}
         <Stepper activeStep={activeStep} orientation="vertical" >
           {steps.map((label, index) => (
             <Step key={label} style={{color: "white", background: '#f5f5f5',borderRadius: '15px', padding:'10px'}}>